adapted for averitec challenge. saving eval_log. refused to answer changed to refuted

+++ src/common/modeling.py
@@ -310,7 +310,7 @@ class LLM(LanguageModel):
                 prompt_length = len(self.encoding.encode(prompt))
                 self.logger.log(
                     orange(f"INFO: Prompt is too long. Length: {prompt_length}. Truncating it hard with factor 0.9."))
-                prompt = prompt[:0.9 * len(prompt)]
+                prompt = prompt[:int(0.9 * len(prompt))]
             with modeling.get_lf_context(gen_temp, max_tokens):
                 while not response and num_attempts < max_attempts:
                     with futures.ThreadPoolExecutor() as executor:
